+++ cloudfit_provider_gcp/__init__.py
@@ -2,7 +2,7 @@
 
 from .provider import GCPProvider
 
-__version__ = "0.1.2"
+__version__ = "0.2.0"
 __author__  = "Chaitanya Krishna Kasaraneni"
 __license__ = "Apache-2.0"